requirements.txt
pillow>=12.2.0
transformers>=4.48.0
torch>=2.6.0
